Fixing fortran control point interface routine. Fortran passes ints by reference not value.

+++ src/ck-cp/controlPointsf.h
@@ -9,8 +9,8 @@ subroutine gotoNextPhase()
 end subroutine
 
 integer function controlPoint(lb, ub)
-     integer :: lb
-integer :: ub
+     integer*4 :: lb
+integer*4 :: ub
 end function
 
 end interface
